refactor(csharp): extract ComplexTypeParser into ComplexTypeSerializingStream.cs
The recursive Databricks SQL type-string parser lived inside
StatementExecutionStatement next to MapDatabricksTypeToArrowType.
Moving it to a sibling internal static class in
ComplexTypeSerializingStream.cs groups the two complementary complex-type
helpers in one file:

  - ComplexTypeSerializingStream:  native Arrow complex array -> JSON string
  - ComplexTypeParser:              SQL type string -> native Arrow type

StatementExecutionStatement.MapDatabricksTypeToArrowType now delegates
to ComplexTypeParser.MapPrimitiveType (for primitives) and
ComplexTypeParser.ParseComplexType (for ARRAY/MAP/STRUCT when the flag
is on). No logic change; just file relocation.

Tests updated to reference ComplexTypeParser. All 15 unit tests + 8
E2E permutations still pass.
